Add Safe deployment callback to safe-web3-lib
