+++ packages/safe-web3-lib/src/contracts/GnosisSafeProxyFactory/GnosisSafeProxyFactoryWeb3Contract.ts
@@ -3,12 +3,14 @@ import { TransactionReceipt } from 'web3-core/types'
 import { ProxyFactory as ProxyFactory_V1_1_1 } from '../../../typechain/src/web3-v1/v1.1.1/proxy_factory'
 import { ProxyFactory as ProxyFactory_V1_3_0 } from '../../../typechain/src/web3-v1/v1.3.0/proxy_factory'
 import { Web3TransactionOptions } from '../../types'
+import { toTxResult } from '../../utils'
 
 export interface CreateProxyProps {
   safeMasterCopyAddress: string
   initializer: string
   saltNonce: number
   options?: Web3TransactionOptions
+  callback?: (txHash: string) => void
 }
 
 class GnosisSafeProxyFactoryWeb3Contract implements GnosisSafeProxyFactoryContract {
@@ -22,7 +24,8 @@ class GnosisSafeProxyFactoryWeb3Contract implements GnosisSafeProxyFactoryContra
     safeMasterCopyAddress,
     initializer,
     saltNonce,
-    options
+    options,
+    callback
   }: CreateProxyProps): Promise<string> {
     if (saltNonce < 0) {
       throw new Error('saltNonce must be greater than 0')
@@ -39,6 +42,12 @@ class GnosisSafeProxyFactoryWeb3Contract implements GnosisSafeProxyFactoryContra
     const txResponse = this.contract.methods
       .createProxyWithNonce(safeMasterCopyAddress, initializer, saltNonce)
       .send(options)
+
+    if (callback) {
+      const txResult = await toTxResult(txResponse)
+      callback(txResult.hash)
+    }
+
     const txResult: TransactionReceipt = await new Promise((resolve, reject) =>
       txResponse.once('receipt', (receipt: TransactionReceipt) => resolve(receipt)).catch(reject)
     )
