+++ packages/safe-web3-lib/src/index.ts
@@ -1,6 +1,5 @@
-import { Web3TransactionResult } from './contracts/GnosisSafe/GnosisSafeContractWeb3'
 import { CreateProxyProps } from './contracts/GnosisSafeProxyFactory/GnosisSafeProxyFactoryWeb3Contract'
-import { Web3TransactionOptions } from './types'
+import { Web3TransactionOptions, Web3TransactionResult } from './types'
 import Web3Adapter, { Web3AdapterConfig } from './Web3Adapter'
 
 export default Web3Adapter
